+++ src/main/java/plus/dragons/createenchantmentindustry/content/contraptions/fluids/experience/ExperienceEffectHandler.java
@@ -2,12 +2,13 @@ package plus.dragons.createenchantmentindustry.content.contraptions.fluids.exper
 
 import com.simibubi.create.content.contraptions.fluids.OpenEndedPipe;
 import com.simibubi.create.foundation.utility.VecHelper;
+
+import io.github.fabricators_of_create.porting_lib.util.FluidStack;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.phys.Vec3;
-import net.minecraftforge.fluids.FluidStack;
 import plus.dragons.createenchantmentindustry.foundation.advancement.CeiAdvancements;
 
 public class ExperienceEffectHandler implements OpenEndedPipe.IEffectHandler {
@@ -29,7 +30,7 @@ public class ExperienceEffectHandler implements OpenEndedPipe.IEffectHandler {
                              pos.getZ() - pipePos.getZ()).scale(0.2);
         var orbPos = VecHelper.getCenterOf(pos);
         ExperienceFluid fluid = (ExperienceFluid) fluidStack.getFluid();
-        int amount = fluidStack.getAmount();
+        int amount = (int) fluidStack.getAmount();
         if (players.isEmpty()) {
             fluid.awardOrDrop(null, level, orbPos, speed, amount);
         } else {
